requirements.txt
PyYAML==5.4
apache-skywalking==0.5.0
